backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
flask-jwt-extended
python-dotenv
psycopg2-binary